feat: #45 컴포넌트 분리
